ecommerce/backend/requirements.txt

# 웹 프레임워크
fastapi                 # FastAPI: REST API 서버 및 웹 요청 처리
uvicorn[standard]       # ASGI 서버: FastAPI 실행용, 빠른 비동기 처리 지원
python-multipart        # Form-data 처리용
email-validator         # 이메일 유효성 검사 (Pydantic[email])

# 데이터베이스
sqlalchemy              # ORM: 파이썬 객체로 DB 조작 가능
pymysql                 # MySQL/MariaDB 드라이버: SQLAlchemy가 DB에 연결할 때 사용
qdrant-client           # 벡터 데이터베이스 클라이언트

# 환경 변수 및 설정
python-dotenv
pydantic-settings       # Pydantic BaseSettings 지원
itsdangerous            # SessionMiddleware 서명/검증

# 인증 및 보안
passlib[bcrypt]         # 비밀번호 해싱
bcrypt<4.0.0           # passlib 72-byte ValueError 호환성 이슈 방지
python-jose[cryptography] # JWT 토큰 생성 및 검증

langchain               # LangChain 핵심
langchain-core          # LangChain 코어
langchain-openai        # OpenAI 연동
langsmith               # LangSmith 트레이싱
langgraph               # LangGraph 워크플로우
transformers            # Guardrail/CLIP 등 모델 파이프라인 import에 필요
# torch/torchvision 은 빌드 시점을 고려해 버전 고정
torch==2.10.0
torchvision==0.25.0
# accelerate는 torch 의존성으로 Docker 빌드 용량 이슈가 있어 EC2 수동 구성에서만 사용
# bitsandbytes는 torch 의존성으로 Docker 빌드 용량 이슈가 있어 EC2 수동 구성에서만 사용

# 임베딩 & 리랭킹
fastembed               # 텍스트 임베딩
flashrank               # 리랭킹 (Reranker)

# 기타 유틸리티
requests                # HTTP 요청
numpy
pandas

# 구글 OAuth2 인증
authlib # OAuth2 클라이언트 및 환경 변수 로딩
